pyproject.toml
[project]
name = "stylegan2-ada-pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.76",
    "click>=8.3.0",
    "gdown>=5.2.0",
    "imageio>=2.37.2",
    "imageio-ffmpeg>=0.6.0",
    "insightface>=0.7.3",
    "matplotlib>=3.10.7",
    "ninja>=1.13.0",
    "numpy>=2.3.4",
    "nvidia-nvimgcodec-cu12>=0.6.1.37",
    "onnxruntime>=1.16.0",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "scikit-image>=0.25.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
